Update with responses and HTML fixes
This updates index.html with candidate responses. The "unknown"
candidate was Jacqui Coyler, and has been noted. (I found this out by
analyzing the MailChimp reports. Perhaps I like it more than I
thought.)

There are also some fixes to the HTML  - mostly eliminating orphan
tags. (I ran it through the validator).

I added a couple of styles to the CSS. One puts a 20-pixel margin
around the page. The other creates a style for the question numbers so
that they can be clicked within the document.